implemented accessor and mutator for date to UTC
