+++ app/Model/Form.php
@@ -5,7 +5,6 @@ namespace App\Model;
 use App\Model\Master\Customer;
 use App\Model\Master\Supplier;
 use App\Model\Master\User;
-use Carbon\Carbon;
 
 class Form extends PointModel
 {
@@ -19,26 +18,6 @@ class Form extends PointModel
         'done',
         'approved',
     ];
-
-    public function setDateAttribute($value)
-    {
-        $this->attributes['date'] = Carbon::parse($value, request()->header('Timezone'))->timezone('UTC')->toDateTimeString();
-    }
-
-    public function getCreatedAtAttribute($value)
-    {
-        return Carbon::parse($value, request()->header('Timezone'))->timezone('UTC')->toDateTimeString();
-    }
-
-    public function getUpdatedAtAttribute($value)
-    {
-        return Carbon::parse($value, request()->header('Timezone'))->timezone('UTC')->toDateTimeString();
-    }
-
-    public function getDateAttribute($value)
-    {
-        return Carbon::parse($value, request()->header('Timezone'))->timezone('UTC')->toDateTimeString();
-    }
 
     public function save(array $options = [])
     {
