+++ app/Model/TransactionModel.php
@@ -3,10 +3,41 @@
 namespace App\Model;
 
 use App\Traits\FormScopes;
+use Carbon\Carbon;
 
 class TransactionModel extends PointModel
 {
     use FormScopes;
+
+    public function setEtaAttribute($value)
+    {
+        $this->attributes['eta'] = Carbon::parse($value, request()->header('Timezone'))->timezone('UTC')->toDateTimeString();
+    }
+
+    public function getEtaAttribute($value)
+    {
+        return Carbon::parse($value, 'UTC')->timezone(request()->header('Timezone'))->toDateTimeString();
+    }
+
+    public function setDueDateAttribute($value)
+    {
+        $this->attributes['due_date'] = Carbon::parse($value, request()->header('Timezone'))->timezone('UTC')->toDateTimeString();
+    }
+
+    public function getDueDateAttribute($value)
+    {
+        return Carbon::parse($value, 'UTC')->timezone(request()->header('Timezone'))->toDateTimeString();
+    }
+
+    public function setRequiredDateAttribute($value)
+    {
+        $this->attributes['required_date'] = Carbon::parse($value, request()->header('Timezone'))->timezone('UTC')->toDateTimeString();
+    }
+
+    public function getRequiredDateAttribute($value)
+    {
+        return Carbon::parse($value, 'UTC')->timezone(request()->header('Timezone'))->toDateTimeString();
+    }
 
     public function edit($data)
     {
